mwaairflow/assets/requirements.txt
apache-airflow==2.0.2
apache-airflow-providers-salesforce==2.0.0
apache-airflow-providers-amazon==1.4.0
apache-airflow-providers-postgres==1.0.2
apache-airflow-providers-mongo==1.0.1
boto3==1.17.112
simplejson==3.17.2
pymongo==3.11.3
mysqlclient==2.0.3
pymssql==2.1.5
smart-open==5.0.0
psycopg2==2.8.6
azure-batch>=8.0.0
azure-cosmos>=3.0.1,<4
azure-datalake-store>=0.0.45
azure-identity>=1.3.1
azure-keyvault>=4.1.0
azure-kusto-data>=0.0.43,<0.1
azure-mgmt-containerinstance>=1.5.0,<2.0
azure-mgmt-datafactory>=1.0.0,<2.0
azure-mgmt-datalake-store>=0.5.0
azure-mgmt-resource>=2.2.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
azure-storage-file==2.1.0
simple-salesforce==1.11.3
